+++ src/main/java/ru/qelezy/app/controllers/DealController.java
@@ -2,12 +2,14 @@ package ru.qelezy.app.controllers;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import ru.qelezy.app.entities.Deal;
 import ru.qelezy.app.repositories.DealRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/deals")
@@ -21,27 +23,44 @@ public class DealController {
     }
 
     @GetMapping("/{id}")
-    public Deal findById(@PathVariable Long id) {
-        return dealRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
+    public ResponseEntity<Object> findById(@PathVariable Long id) {
+        Optional<Deal> dealOptional = dealRepository.findById(id);
+        if (dealOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(dealOptional.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void add(@RequestBody Deal deal) {
+    public ResponseEntity<Object> add(@RequestBody Deal deal) {
         dealRepository.save(deal);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void update(@PathVariable Long id, @RequestBody @NotNull Deal newDeal) {
-        Deal existedDeal = dealRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
-        newDeal.setId(id);
-        dealRepository.save(newDeal);
+    public ResponseEntity<Object> update(@PathVariable Long id, @RequestBody @NotNull Deal newDeal) {
+        Optional<Deal> dealOptional = dealRepository.findById(id);
+        if (dealOptional.isPresent()) {
+            newDeal.setId(id);
+            dealRepository.save(newDeal);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable Long id) {
-        dealRepository.deleteById(id);
+    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+        Optional<Deal> dealOptional = dealRepository.findById(id);
+        if (dealOptional.isPresent()) {
+            dealRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 }
