+++ src/main/java/ru/qelezy/app/controllers/ExchangeController.java
@@ -2,12 +2,14 @@ package ru.qelezy.app.controllers;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.qelezy.app.entities.Exchange;
 import ru.qelezy.app.repositories.ExchangeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/exchanges")
@@ -21,27 +23,44 @@ public class ExchangeController {
     }
 
     @GetMapping("/{id}")
-    public Exchange findById(@PathVariable Long id) {
-        return exchangeRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
+    public ResponseEntity<Object> findById(@PathVariable Long id) {
+        Optional<Exchange> exchangeOptional = exchangeRepository.findById(id);
+        if (exchangeOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(exchangeOptional.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void add(@RequestBody Exchange exchange) {
+    public ResponseEntity<Object> add(@RequestBody Exchange exchange) {
         exchangeRepository.save(exchange);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void update(@PathVariable Long id, @RequestBody @NotNull Exchange newExchange) {
-        Exchange existedExchange = exchangeRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
-        newExchange.setId(id);
-        exchangeRepository.save(newExchange);
+    public ResponseEntity<Object> update(@PathVariable Long id, @RequestBody @NotNull Exchange newExchange) {
+        Optional<Exchange> exchangeOptional = exchangeRepository.findById(id);
+        if (exchangeOptional.isPresent()) {
+            newExchange.setId(id);
+            exchangeRepository.save(newExchange);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable Long id) {
-        exchangeRepository.deleteById(id);
+    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+        Optional<Exchange> exchangeOptional = exchangeRepository.findById(id);
+        if (exchangeOptional.isPresent()) {
+            exchangeRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 }
