+++ src/main/java/ru/qelezy/app/controllers/PortfolioController.java
@@ -2,12 +2,14 @@ package ru.qelezy.app.controllers;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.qelezy.app.entities.Portfolio;
 import ru.qelezy.app.repositories.PortfolioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/portfolios")
@@ -21,27 +23,44 @@ public class PortfolioController {
     }
 
     @GetMapping("/{id}")
-    public Portfolio findById(@PathVariable Long id) {
-        return portfolioRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
+    public ResponseEntity<Object> findById(@PathVariable Long id) {
+        Optional<Portfolio> portfolioOptional = portfolioRepository.findById(id);
+        if (portfolioOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(portfolioOptional.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void add(@RequestBody Portfolio portfolio) {
+    public ResponseEntity<Object> add(@RequestBody Portfolio portfolio) {
         portfolioRepository.save(portfolio);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void update(@PathVariable Long id, @RequestBody @NotNull Portfolio newPortfolio) {
-        Portfolio existedPortfolio = portfolioRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
-        newPortfolio.setId(id);
-        portfolioRepository.save(newPortfolio);
+    public ResponseEntity<Object> update(@PathVariable Long id, @RequestBody @NotNull Portfolio newPortfolio) {
+        Optional<Portfolio> portfolioOptional = portfolioRepository.findById(id);
+        if (portfolioOptional.isPresent()) {
+            newPortfolio.setId(id);
+            portfolioRepository.save(newPortfolio);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable Long id) {
-        portfolioRepository.deleteById(id);
+    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+        Optional<Portfolio> portfolioOptional = portfolioRepository.findById(id);
+        if (portfolioOptional.isPresent()) {
+            portfolioRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 }
