+++ src/main/java/ru/qelezy/app/controllers/TrendController.java
@@ -2,12 +2,14 @@ package ru.qelezy.app.controllers;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.qelezy.app.entities.Trend;
 import ru.qelezy.app.repositories.TrendRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/trends")
@@ -21,27 +23,44 @@ public class TrendController {
     }
 
     @GetMapping("/{id}")
-    public Trend findById(@PathVariable Long id) {
-        return trendRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
+    public ResponseEntity<Object> findById(@PathVariable Long id) {
+        Optional<Trend> trendOptional = trendRepository.findById(id);
+        if (trendOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.OK).body(trendOptional.get());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void add(@RequestBody Trend trend) {
+    public ResponseEntity<Object> add(@RequestBody Trend trend) {
         trendRepository.save(trend);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void update(@PathVariable Long id, @RequestBody @NotNull Trend newTrend) {
-        Trend existedTrend = trendRepository.findById(id).orElseThrow(() -> new NullPointerException("Запись с id: " + id + "не найдена"));
-        newTrend.setId(id);
-        trendRepository.save(newTrend);
+    public ResponseEntity<Object> update(@PathVariable Long id, @RequestBody @NotNull Trend newTrend) {
+        Optional<Trend> trendOptional = trendRepository.findById(id);
+        if (trendOptional.isPresent()) {
+            newTrend.setId(id);
+            trendRepository.save(newTrend);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteById(@PathVariable Long id) {
-        trendRepository.deleteById(id);
+    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+        Optional<Trend> trendOptional = trendRepository.findById(id);
+        if (trendOptional.isPresent()) {
+            trendRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Запись с id: " + id + " не найдена");
+        }
     }
 }
